Simplify dev/docs by moving from example app to storybook.
