+++ .storybook/config.js
@@ -1,0 +1,4 @@
+import { configure } from "@storybook/react";
+
+// automatically import all files ending in *.stories.js
+configure(require.context("../src", true, /\.stories\.[tj]sx?$/), module);
