+++ .storybook/webpack.config.js
@@ -1,0 +1,25 @@
+module.exports = ({ config }) => {
+  config.module.rules.push(
+    {
+      test: /\.(ts|tsx)$/,
+      use: [
+        {
+          loader: require.resolve("awesome-typescript-loader")
+        }
+      ]
+    },
+    {
+      test: /\.s[ac]ss$/i,
+      use: [
+        // Creates `style` nodes from JS strings
+        "style-loader",
+        // Translates CSS into CommonJS
+        "css-loader",
+        // Compiles Sass to CSS
+        "sass-loader"
+      ]
+    }
+  );
+  config.resolve.extensions.push(".ts", ".tsx");
+  return config;
+};
