--- example/.npmignore
@@ -1,3 +1,0 @@
-node_modules
-.cache
-dist
