--- example/index.css
@@ -1,9 +1,0 @@
-body {
-  margin: 0;
-  padding: 0;
-  font-family: sans-serif;
-
-  width: 600px;
-  margin-left: auto;
-  margin-right: auto;
-}
