--- example/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "react-datetime-example",
-  "homepage": "https://NateRadebaugh.github.io/react-datetime",
-  "version": "0.0.0",
-  "main": "index.js",
-  "engines": {
-    "node": "10.x"
-  },
-  "license": "MIT",
-  "private": true,
-  "scripts": {
-    "now-build": "cd .. && yarn install && yarn build && cd example && yarn install && yarn build",
-    "dev": "parcel index.html",
-    "start": "yarn dev",
-    "build": "parcel build index.html"
-  },
-  "dependencies": {
-    "date-fns": "^2.7.0",
-    "react": "^16.11.0",
-    "react-app-polyfill": "^1.0.4",
-    "react-dom": "^16.11.0"
-  },
-  "alias": {
-    "react": "../node_modules/react",
-    "react-dom": "../node_modules/react-dom"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.7.2",
-    "@types/react": "^16.9.11",
-    "@types/react-dom": "^16.9.4",
-    "parcel": "^1.12.4",
-    "typescript": "^3.7.2"
-  }
-}
