--- example/public/manifest.json
@@ -1,8 +1,0 @@
-{
-  "short_name": "react-datetime",
-  "name": "react-datetime",
-  "start_url": "./index.html",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
