+++ now.json
@@ -1,6 +1,5 @@
 {
   "name": "@nateradebaugh/react-datetime",
   "version": 2,
-  "builds": [{ "src": "example/package.json", "use": "@now/static-build" }],
-  "routes": [{ "src": "/(.*)", "dest": "/example/$1" }]
+  "builds": [{ "src": "package.json", "use": "@now/static-build" }]
 }
